Implement StairCaseEnvelope, code cleanup.
